feat: Enhance stack convert and fix lint issues
- Add PlaceholderFunction base type for PostMerge functions to reduce
  code duplication (store, terraform functions)
- Consolidate store.go and terraform.go into placeholder_functions.go
- Fix staticcheck QF1008: remove redundant embedded field from selector
- Add new sentinel errors for stack loader operations
- Enhance stack convert with multi-document support
- Update remark-stack-example converter for HCL block syntax

+++ pkg/function/random.go
@@ -108,7 +108,8 @@ func generateRandomNumber(min, max int) (int, error) {
 	}
 
 	// Generate cryptographically secure random number in range [min, max].
-	rangeSize := int64(max - min + 1)
+	// Cast to int64 before arithmetic to prevent overflow with large ranges.
+	rangeSize := int64(max) - int64(min) + 1
 	n, err := rand.Int(rand.Reader, big.NewInt(rangeSize))
 	if err != nil {
 		return 0, fmt.Errorf("failed to generate random number: %w", err)
